+++ packages/voice/src/models/v2/validate-svaml-request/index.ts
@@ -1,0 +1,1 @@
+export type { ValidateSvamlRequest, ValidationTypeEnum } from './validate-svaml-request';